% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{Misc.AppendRData}
\alias{Misc.AppendRData}
\title{Append to current RData file}
\usage{
Misc.AppendRData(..., list = character(), file)
}
\arguments{
\item{...}{objects}

\item{file}{RData file}
}
\description{
Append to current RData file
}
