% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{Misc.InstallCommon}
\alias{Misc.InstallCommon}
\title{Install commonly used packages
To get a newly installed R version ready}
\usage{
Misc.InstallCommon(Packages = c("caret", "randomForest", "e1071", "ggplot2",
  "plotly", "shiny", "rsconnect", "xlsx", "nnet", "sqldf", "devtools",
  "roxygen2"))
}
\description{
Install commonly used packages
To get a newly installed R version ready
}
