% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{Misc.PrintLinearEquation}
\alias{Misc.PrintLinearEquation}
\title{Print Linear Equation
Print the linear equation from a glm or lm object}
\usage{
Misc.PrintLinearEquation(model, round = 3)
}
\arguments{
\item{model}{a glm or lm object}

\item{round}{rounding for parameters}
}
\description{
Print Linear Equation
Print the linear equation from a glm or lm object
}
