% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Misc.R
\name{Misc.setwdClipboard}
\alias{Misc.setwdClipboard}
\title{Set current working directory to the one copied on clipboard}
\usage{
Misc.setwdClipboard()
}
\description{
Set current working directory to the one copied on clipboard
}
