% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiCol.R
\name{Multicol.IndividualTest}
\alias{Multicol.IndividualTest}
\title{Individual Multicollinearity Diagnostic Measures}
\usage{
Multicol.IndividualTest(x, y, method = NULL, na.rm = TRUE, corr = FALSE,
  vif = 10, tol = 0.1, conf = 0.95, cvif = 10, leamer = 0.1,
  all = FALSE)
}
\description{
methods are VIF, CVIF, TOL, Leamer, Wi, Fi, Klein
}
