% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.ClusterElbow}
\alias{Plot.ClusterElbow}
\title{Plot the Elbow curve for clustering}
\usage{
Plot.ClusterElbow(Data, kmax = 8)
}
\description{
use hierachical clustering to plot the within-group variance against number of clusters
}
