% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.PublishPlot}
\alias{Plot.PublishPlot}
\title{Save Publication-quality plot}
\usage{
Plot.PublishPlot(plot, res = 600, height = 2000, width = 4000,
  dir = getwd(), filename = "Plot Output")
}
\arguments{
\item{plot}{any plot or ggplot object}

\item{res}{resolution (dpi)}

\item{height}{height (in pixels)}

\item{width}{width (in pixels)}

\item{dir}{saving directory}

\item{filename}{file name}
}
\description{
Savev a plot in a high-quality jpeg file for publication
}
