% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{Stats.StandardANOVAPractice}
\alias{Stats.StandardANOVAPractice}
\title{Workflow of standard Analysis-of-variabce practice
We test for difference in risk factors between the cohort groups. If the data passes both normality check and homogeneity check, we used one-way ANOVA test. If the data passes normality check but fails homogeneity check, we used Welch's one-way test. If the data fails normality check, we used Kruskal-Wallis test. Then, we used Tukey test to discover the pairs of cohorts that have significant difference.}
\usage{
Stats.StandardANOVAPractice(y, x, observe = TRUE, Thres.Normality = 0.05,
  Thres.Homogeneity = 0.05, Thres.Decision = 0.05)
}
\arguments{
\item{y}{vector of Groups, must be a factor}

\item{x}{matrix or data frame of measures, should include column names}

\item{observe}{if TRUE, wait for user to press enter through each step}
}
\description{
Workflow of standard Analysis-of-variabce practice
We test for difference in risk factors between the cohort groups. If the data passes both normality check and homogeneity check, we used one-way ANOVA test. If the data passes normality check but fails homogeneity check, we used Welch's one-way test. If the data fails normality check, we used Kruskal-Wallis test. Then, we used Tukey test to discover the pairs of cohorts that have significant difference.
}
