% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.PublishPlot}
\alias{Plot.PublishPlot}
\title{Save Publication-quality plot}
\usage{
Plot.PublishPlot(plot, res = 400, height = 1500, width = 3000,
  dir = getwd(), filename = "Plot Output")
}
\arguments{
\item{plot}{any plot or ggplot object}

\item{res}{resolution}

\item{height}{height}

\item{width}{width}

\item{dir}{saving directory}

\item{filename}{file name}
}
\description{
Save Publication-quality plot
}
