% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{Plot.VisualizeSupervise}
\alias{Plot.VisualizeSupervise}
\title{Visualize High-dimensional data in a 3D scatter plot
Using PCA to visualize a classification data at 2D or 3D}
\usage{
Plot.VisualizeSupervise(formula, data, dim = NULL)
}
\description{
Visualize High-dimensional data in a 3D scatter plot
Using PCA to visualize a classification data at 2D or 3D
}
